requirements.txt
fastapi
uvicorn[standard]
matplotlib
plotly
openai
requests
websockets
pandas
seaborn
python-dotenv
numpy
